+++ docs/CHANGELOG.md
@@ -23,6 +23,8 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Specialised UI prefab format ([#786])
 * Add generation of normals/tangents in GLTF ([#784])
 * Derive `PartialEq` for `SpriteSheet` ([#789])
+* Add basic scene prefab ([#791])
+* Improve ergonomics of examples ([#XXX])
 
 ### Changed
 * UI systems will now never overwrite your local `UiTransform` values ([#774])
@@ -66,7 +68,9 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#784]: https://github.com/amethyst/amethyst/pull/784
 [#785]: https://github.com/amethyst/amethyst/pull/785
 [#786]: https://github.com/amethyst/amethyst/pull/786
+[#791]: https://github.com/amethyst/amethyst/pull/791
 [#789]: https://github.com/amethyst/amethyst/pull/789
+[#XXX]: https://github.com/amethyst/amethyst/pull/XXX
 
 ## [0.7.0] - 2018-05
 ### Added